Change to the better namespace and rename the class/enum.
