+++ osu.Game.Rulesets.Karaoke/Objects/Note.cs
@@ -8,6 +8,7 @@ using osu.Game.Rulesets.Judgements;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.Judgements;
 using osu.Game.Rulesets.Karaoke.Objects.Types;
+using osu.Game.Rulesets.Karaoke.Objects.Workings;
 using osu.Game.Rulesets.Karaoke.Scoring;
 using osu.Game.Rulesets.Objects.Types;
 using osu.Game.Rulesets.Scoring;
@@ -111,7 +112,7 @@ namespace osu.Game.Rulesets.Karaoke.Objects
                     return;
 
                 // should trying to reload the reference lyric.
-                Validator.Invalidate(NoteInvalidation.ReferenceLyric);
+                Validator.Invalidate(NoteWorkingProperty.ReferenceLyric);
             }
         }
