+++ osu.Game.Rulesets.Karaoke/Objects/Note_Working.cs
@@ -17,7 +17,7 @@ namespace osu.Game.Rulesets.Karaoke.Objects;
 public partial class Note
 {
     [JsonIgnore]
-    public HitObjectValidator<NoteInvalidation> Validator { get; } = new();
+    public HitObjectWorkingPropertyValidator<NoteWorkingProperty> Validator { get; } = new();
 
     [JsonIgnore]
     public readonly Bindable<int?> PageIndexBindable = new();
